+++ src/main/kotlin/br/com/clmgni/bytebank2/main.kt
@@ -1,32 +1,5 @@
-import br.com.clmgni.bytebank2.modelo.Endereco
+import br.com.clmgni.bytebank2.teste.testaAny
 
 fun main() {
-    println("Iniciando o Principal!")
-
-    val endereco = Endereco(
-        logradouro = "Rua vergueiro",
-        complemento = "Alura",
-        cep = "00000-0700"
-    )
-    val enderecoNovo = Endereco(
-        logradouro = "Rua vergueiro",
-        complemento = "Alura",
-        cep = "00000-0700"
-    )
-
-    println(endereco.equals(enderecoNovo))
-
-    println(endereco.hashCode())
-    println(enderecoNovo.hashCode())
-
-    println(endereco)
-    println(enderecoNovo)
-
-    println("${endereco.javaClass}@${
-        Integer.toHexString(endereco.hashCode())}")
-}
-
-fun imprime(valor: Any): Any {
-    println(valor)
-    return valor
+    testaAny()
 }
