+++ src/main/kotlin/br/com/clmgni/bytebank2/modelo/Analista.kt
@@ -1,7 +1,5 @@
 package br.com.clmgni.bytebank2.modelo
 
-import br.com.clmgni.bytebank2.modelo.Funcionario
-
 class Analista(
     nome: String,
     cpf: String,
