+++ src/main/kotlin/br/com/clmgni/bytebank2/teste/testaComportamentosConta.kt
@@ -1,3 +1,5 @@
+package br.com.clmgni.bytebank2.teste
+
 import br.com.clmgni.bytebank2.modelo.Cliente
 import br.com.clmgni.bytebank2.modelo.ContaCorrente
 import br.com.clmgni.bytebank2.modelo.ContaPoupanca
