+++ src/main/kotlin/br/com/clmgni/bytebank2/teste/testaFuncionarios.kt
@@ -1,7 +1,6 @@
-import br.com.clmgni.bytebank2.modelo.Analista
-import br.com.clmgni.bytebank2.modelo.CalculadoraBonificacao
-import br.com.clmgni.bytebank2.modelo.Diretor
-import br.com.clmgni.bytebank2.modelo.Gerente
+package br.com.clmgni.bytebank2.teste
+
+import br.com.clmgni.bytebank2.modelo.*
 
 fun testaFuncionarios() {
     val alex = Analista(
